// chpqa-app/src/app/tasks/scheme-registration/models/scheme-registration-path.model.ts
export enum SchemeRegistartiondPath {
  BASE_PATH = '/scheme-registration',
  BEFORE_REGISTRATION_STARTS = 'before-registration-starts',
  ENTER_SCHEME_NAME = 'enter-scheme-name',
  SELECT_SIC_CODE = 'select-sic-code',
  SELECT_ECONOMIC_SECTOR = 'select-economic-sector',
  SELECT_ECONOMIC_SUB_SECTOR = 'select-economic-sub-sector',
  SEARCH_SITE_ADDRESS = 'search-site-address',
  SELECT_SITE_ADDRESS = 'select-site-address',
  ADD_SITE_ADDRESS = 'add-site-address',
  CONFIRM_SITE_ADDRESS = 'confirm-site-address',
  CONFIRM_SITE_CONTACT = 'confirm-site-contact',
  ADD_SITE_CONTACT = 'add-site-contact',
  ADD_NEW_SITE_CONTACT = 'add-new-site-contact',
  SITE_CONTACT_DETAILS = 'site-contact-details',
  NEW_SITE_CONTACT_DETAILS = 'new-site-contact-details',
  EXISTING_SITE_CONTACT_DETAILS = 'existing-site-contact-details',
  CONFIRMATION = 'confirmation',
  SUMMARY = 'summary',
}